Consolidate source into a single directory.
Also, get unit tests running with console reporter.
Fix up error reporting in console reporter.
