+++ src/Acc.js
@@ -6,7 +6,7 @@
 // Convenience.
 var {observe, unobserve} = Object;
 
-import {Observer} from '../Utl/Obs';
+import {Observer} from './Obs';
 
 // AccessNotifier
 // --------------
@@ -33,26 +33,26 @@ function makeAccessController(rerun) {
   // `names` of null means to watch properties of any name.
   // It is built by calls to `notifyAccess`, invoked through `accessNotifier`.
   var accesses = new Map();
-  
+
   function unobserve() {
     accesses.forEach(({observer}) => observer.unobserve());
   }
-  
+
   function observe() {
     accesses.forEach(({observer}) => observer.observe(['update', 'add', 'delete']));
   }
-  
+
   // Start capturing accessed dependencies.
   function capture() {
     accesses.clear();
     accessNotifier.push(notifyAccess);
   }
-  
+
   // Stop capturing accessed dependencies.
   function uncapture() {
     accessNotifier.pop();
   }
-  
+
   function start() {
     unobserve();
     capture();
@@ -66,7 +66,7 @@ function makeAccessController(rerun) {
   // `notifyAccess` is the callback invoked by upwardables when a property is accessed.
   // It records the access in the `accesses` map.
   function notifyAccess({object, name}) {
-    
+
     // Create an observer for changes in properties accessed during execution of this function.
     function makeAccessedObserver() {
       return Observer(object, function(changes) {
@@ -77,7 +77,7 @@ function makeAccessController(rerun) {
         });
       });
     }
-    
+
     // Make a new entry in the access table, containing initial property name if any
     // and observer for properties accessed on the object.
     function makeAccessEntry() {
@@ -86,19 +86,19 @@ function makeAccessController(rerun) {
         observer: makeAccessedObserver()
       });
     }
-    
+
     // If properties on this object are already being watched, there is already an entry
     // in the access table for it. Add a new property name to the existing entry.
     function setAccessEntry() {
       if (name && accessEntry.names) accessEntry.names.push(name);
       else accessEntry.names = null;
     }
-    
+
     var accessEntry = accesses.get(object);
     if (accessEntry) setAccessEntry();
     else makeAccessEntry();
   }
-  
+
   return {start, stop};
 }
