+++ src/Fet.js
@@ -2,10 +2,16 @@
 //
 // Utilities for fetch.
 
+
 function fetchJson(...args) {
   return fetch(...args) . then(response => response.json());
 }
 
+function fetchText(...args) {
+  return fetch(...args) . then(response => response.text());
+}
+
 export {
-  fetchJson
+  fetchJson,
+  fetchText
 };
