+++ src/Srt.js
@@ -3,18 +3,14 @@
 
 import {makeUpwardableFunction} from './Fun';
 import makeUpwardableObject from './Obj';
-import {copyOntoArray} from '../Utl/Obj';
-import {makeSortfunc} from '../Utl/Utl';
+import {copyOntoArray} from './Out';
+import {makeSortfunc} from './Utl';
 
 export default makeUpwardableFunction(function *UpSort(run) {
   var r = [];
-  
+
   while (true) {
     var [a, f, desc] = yield r;
     copyOntoArray(r, a.slice().sort(makeSortfunc(f, desc)))
   }
 });
-
-
-
-
