+++ src/Str.js
@@ -1,19 +1,19 @@
 // String utilities
 // ----------------
 
-import {upwardConfig} from '../Cfg';
+import {upwardConfig} from './Cfg';
 
 // `my-class` => `myClass`
 function camelize(str) {
   return str.replace(
-      /[-_]+([a-z])/g, 
+      /[-_]+([a-z])/g,
     (_, letter) => letter.toUpperCase());
 }
 
 // `myClass` => `my-class`
 function dasherize(str) {
   return str.replace(
-      /([a-z])([A-Z])/g, 
+      /([a-z])([A-Z])/g,
     (_, let1, let2) => `${let1}-${let2.toLowerCase()}`
   );
 }
