+++ src/Up.js
@@ -3,18 +3,16 @@
 
 // Re-export interfaces so clients can import from this single module.
 
-import U       from './Upw/Obj';
-import T       from './Upw/Txt';
-import E       from './Upw/Elt';
-import F       from './Upw/Tem';
-import C       from './Upw/Fun';
+import              './Arr';
+import UpStyle from './Css';
+import UpCount from './Cnt';
+import E       from './Elt';
+import C       from './Fun';
+import U       from './Obj';
+import F       from './Tem';
+import T       from './Txt';
 
-import UpCount from './Upw/Cnt';
-import UpStyle from './Upw/Css';
-
-import              './Upw/Arr';
 
 export {U, T, E, F, C, UpCount, UpStyle};
 
-//export {P, H1, H2, H3, H4, H5, H6, B, I, LI, LABEL, A, BUTTON} from './Upw/Tag';
-export * from './Upw/Tag';
+export * from './Tag';
