+++ tst/index.html
@@ -1,6 +1,6 @@
 <!DOCTYPE HTML>
 
-<!-- 
+<!--
      HTML for Upward test runner.
      Modify run.js to configure tests.
 -->
@@ -10,18 +10,25 @@
   <head>
     <meta http-equiv="Content-Type" content="text/html; charset=UTF-8"/>
     <title>Upward Test Runner</title>
-    
+
     <link rel="icon" type="image/png" href="../ass/img/upward.png"/>
 
     <!-- testing libraries -->
-    <script src="http://chaijs.com/chai.js"></script>  
+    <script src="http://chaijs.com/chai.js"></script>
     <script src="http://cdnjs.cloudflare.com/ajax/libs/sinon.js/1.7.3/sinon.js"></script>
-    
-    <!-- traceur -->
-    <script src="https://google.github.io/traceur-compiler/bin/traceur.js"></script>
-    <script src="https://google.github.io/traceur-compiler/src/bootstrap.js"></script>
-    <script>traceur.options.experimental = true;</script>
-    
+
+    <!-- Babel transpiler -->
+    <script src="../node_modules/babel-core/browser.js"></script>
+
+    <!-- Module loader -->
+    <script src="../node_modules/es6-module-loader/dist/es6-module-loader-dev.js"></script>
+
+    <!-- Babel polyfill, needed for generators. -->
+    <script src="../node_modules/babel-core/browser-polyfill.js"></script>
+
+    <!-- Configure es6-module loader. -->
+    <script src="../es6-module-loader-config.js"></script>
+
     <!-- livereload -->
     <script>
       document.write('<' + 'script src="http://' + (location.host || 'localhost').split(':')[0] + ':35729/livereload.js?snipver=1"></' + 'script>')
@@ -29,7 +36,7 @@
   </head>
 
   <body>
-    <script type="module" src="./run.js"></script>
+    <script>System.import('./runner');</script>
   </body>
 
 </html>
